requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
colorama==0.4.4
Django==3.2.9
django-crispy-forms==1.13.0
django-environ==0.10.0
idna==3.3
Pillow==8.4.0
pytz==2021.3
qrcode==7.3.1
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
